+++ packages/expect-type/src/__tests__/index.test.ts
@@ -1,19 +1,61 @@
 import {expectTypeOf} from '..'
 
-test('Type-check object references', () => {
+test('Check that two objects have equivalent types with `.toEqualTypeOf`', () => {
   expectTypeOf({a: 1}).toEqualTypeOf({a: 1})
-  expectTypeOf({a: 1, b: 1}).toMatchTypeOf({a: 1})
+})
+
+test('`.toEqualTypeOf` succeeds for objects with different values, but the same type', () => {
   expectTypeOf({a: 1}).toEqualTypeOf({a: 2})
 })
 
-test('Assertions can be inverted', () => {
+test("When there's no instance/runtime variable for the expected type, you can use generics", () => {
+  expectTypeOf({a: 1}).toEqualTypeOf<{a: number}>()
+})
+
+test('`.toEqualTypeOf` fails on extra properties', () => {
+  // @ts-expect-error
+  expectTypeOf({a: 1, b: 1}).toEqualTypeOf({a: 1})
+})
+
+test('To allow for extra properties, use `.toMatchTypeOf`. This checks that an object "matches" a type. This is similar to jest\'s `.toMatchObject`', () => {
+  expectTypeOf({a: 1, b: 1}).toMatchTypeOf({a: 1})
+})
+
+test('Another example of the difference between `.toMatchTypeOf` and `.toEqualTypeOf`, using generics. `.toMatchTypeOf` can be used for "is-a" relationships', () => {
+  type Fruit = {type: 'Fruit'; edible: boolean}
+  type Apple = {type: 'Fruit'; name: 'Apple'; edible: true}
+
+  expectTypeOf<Apple>().toMatchTypeOf<Fruit>()
+
+  // @ts-expect-error
+  expectTypeOf<Fruit>().toMatchTypeOf<Apple>()
+
+  // @ts-expect-error
+  expectTypeOf<Apple>().toEqualTypeOf<Fruit>()
+})
+
+test('Assertions can be inverted with `.not`', () => {
   expectTypeOf({a: 1}).not.toMatchTypeOf({b: 1})
+})
+
+test('`.not` can be easier than relying on `// @ts-expect-error`', () => {
+  type Fruit = {type: 'Fruit'; edible: boolean}
+  type Apple = {type: 'Fruit'; name: 'Apple'; edible: true}
+
+  expectTypeOf<Apple>().toMatchTypeOf<Fruit>()
+
+  expectTypeOf<Fruit>().not.toMatchTypeOf<Apple>()
+  expectTypeOf<Apple>().not.toEqualTypeOf<Fruit>()
 })
 
 test('Catch any/unknown/never types', () => {
   expectTypeOf<unknown>().toBeUnknown()
   expectTypeOf<any>().toBeAny()
   expectTypeOf<never>().toBeNever()
+})
+
+test('`.toEqualTypeOf` distinguishes between deeply-nested `any` and `unknown` properties', () => {
+  expectTypeOf<{deeply: {nested: any}}>().not.toEqualTypeOf<{deeply: {nested: unknown}}>()
 })
 
 test('Test for basic javascript types', () => {
@@ -41,7 +83,7 @@ test('Nullable types', () => {
   expectTypeOf<1 | undefined | null>().toBeNullable()
 })
 
-test('Assertions can be inverted with `.not`', () => {
+test('More `.not` examples', () => {
   expectTypeOf(1).not.toBeUnknown()
   expectTypeOf(1).not.toBeAny()
   expectTypeOf(1).not.toBeNever()
@@ -80,6 +122,19 @@ test('Assert on function parameters (using `.parameter(n)` or `.parameters`) and
   const twoArgFunc = (a: number, b: string) => ({a, b})
 
   expectTypeOf(twoArgFunc).parameters.toEqualTypeOf<[number, string]>()
+})
+
+test('Assert on constructor parameters', () => {
+  expectTypeOf(Date).toBeConstructibleWith('1970')
+  expectTypeOf(Date).toBeConstructibleWith(0)
+  expectTypeOf(Date).toBeConstructibleWith(new Date())
+  expectTypeOf(Date).toBeConstructibleWith()
+
+  expectTypeOf(Date).constructorParameters.toEqualTypeOf<[] | [string | number | Date]>()
+})
+
+test('Class instance types', () => {
+  expectTypeOf(Date).instance.toHaveProperty('toISOString')
 })
 
 test('Promise resolution types can be checked with `.resolves`', () => {